api: queue: add separate header for type definitions
Enable easier function inlining by moving type definitions into a separate
header file.

Tested-by: Ashwin Sekhar T K <asekhar@marvell.com>
