+++ include/Makefile.am
@@ -38,6 +38,7 @@ odpapiinclude_HEADERS = \
 	odp/api/protocols.h \
 	odp/api/pool.h \
 	odp/api/queue.h \
+	odp/api/queue_types.h \
 	odp/api/random.h \
 	odp/api/reassembly.h \
 	odp/api/rwlock.h \
@@ -145,6 +146,7 @@ odpapiabidefaultinclude_HEADERS = \
 	odp/api/abi-default/packet_io.h \
 	odp/api/abi-default/pool.h \
 	odp/api/abi-default/queue.h \
+	odp/api/abi-default/queue_types.h \
 	odp/api/abi-default/rwlock.h \
 	odp/api/abi-default/rwlock_recursive.h \
 	odp/api/abi-default/schedule.h \
@@ -192,6 +194,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/arm32-linux/odp/api/abi/packet_io.h \
 	odp/arch/arm32-linux/odp/api/abi/pool.h \
 	odp/arch/arm32-linux/odp/api/abi/queue.h \
+	odp/arch/arm32-linux/odp/api/abi/queue_types.h \
 	odp/arch/arm32-linux/odp/api/abi/rwlock.h \
 	odp/arch/arm32-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/arm32-linux/odp/api/abi/schedule.h \
@@ -235,6 +238,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/arm64-linux/odp/api/abi/packet_io.h \
 	odp/arch/arm64-linux/odp/api/abi/pool.h \
 	odp/arch/arm64-linux/odp/api/abi/queue.h \
+	odp/arch/arm64-linux/odp/api/abi/queue_types.h \
 	odp/arch/arm64-linux/odp/api/abi/rwlock.h \
 	odp/arch/arm64-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/arm64-linux/odp/api/abi/schedule.h \
@@ -278,6 +282,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/default-linux/odp/api/abi/packet_io.h \
 	odp/arch/default-linux/odp/api/abi/pool.h \
 	odp/arch/default-linux/odp/api/abi/queue.h \
+	odp/arch/default-linux/odp/api/abi/queue_types.h \
 	odp/arch/default-linux/odp/api/abi/rwlock.h \
 	odp/arch/default-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/default-linux/odp/api/abi/schedule.h \
@@ -321,6 +326,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/mips64-linux/odp/api/abi/packet_io.h \
 	odp/arch/mips64-linux/odp/api/abi/pool.h \
 	odp/arch/mips64-linux/odp/api/abi/queue.h \
+	odp/arch/mips64-linux/odp/api/abi/queue_types.h \
 	odp/arch/mips64-linux/odp/api/abi/rwlock.h \
 	odp/arch/mips64-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/mips64-linux/odp/api/abi/schedule.h \
@@ -364,6 +370,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/power64-linux/odp/api/abi/packet_io.h \
 	odp/arch/power64-linux/odp/api/abi/pool.h \
 	odp/arch/power64-linux/odp/api/abi/queue.h \
+	odp/arch/power64-linux/odp/api/abi/queue_types.h \
 	odp/arch/power64-linux/odp/api/abi/rwlock.h \
 	odp/arch/power64-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/power64-linux/odp/api/abi/schedule.h \
@@ -407,6 +414,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/x86_32-linux/odp/api/abi/packet_io.h \
 	odp/arch/x86_32-linux/odp/api/abi/pool.h \
 	odp/arch/x86_32-linux/odp/api/abi/queue.h \
+	odp/arch/x86_32-linux/odp/api/abi/queue_types.h \
 	odp/arch/x86_32-linux/odp/api/abi/rwlock.h \
 	odp/arch/x86_32-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/x86_32-linux/odp/api/abi/schedule.h \
@@ -450,6 +458,7 @@ odpapiabiarchinclude_HEADERS = \
 	odp/arch/x86_64-linux/odp/api/abi/packet_io.h \
 	odp/arch/x86_64-linux/odp/api/abi/pool.h \
 	odp/arch/x86_64-linux/odp/api/abi/queue.h \
+	odp/arch/x86_64-linux/odp/api/abi/queue_types.h \
 	odp/arch/x86_64-linux/odp/api/abi/rwlock.h \
 	odp/arch/x86_64-linux/odp/api/abi/rwlock_recursive.h \
 	odp/arch/x86_64-linux/odp/api/abi/schedule.h \
