+++ include/odp/api/packet.h
@@ -22,7 +22,6 @@ extern "C" {
 #include <odp/api/abi/packet_io.h>
 #include <odp/api/abi/packet_types.h>
 #include <odp/api/abi/packet.h>
-#include <odp/api/abi/queue.h>
 #include <odp/api/abi/buffer.h>
 #include <odp/api/abi/pool.h>
