+++ include/odp/api/packet_io.h
@@ -22,7 +22,7 @@ extern "C" {
 #include <odp/api/abi/classification.h>
 #include <odp/api/abi/packet_types.h>
 #include <odp/api/abi/packet_io.h>
-#include <odp/api/abi/queue.h>
+#include <odp/api/abi/queue_types.h>
 
 /** @ingroup odp_packet_io
  *  @{
