+++ include/odp/api/queue_types.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2015-2018, Linaro Limited
+/* Copyright (c) 2021, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -10,15 +10,16 @@
  * ODP queue
  */
 
-#ifndef ODP_API_ABI_QUEUE_H_
-#define ODP_API_ABI_QUEUE_H_
+#ifndef ODP_API_QUEUE_TYPES_H_
+#define ODP_API_QUEUE_TYPES_H_
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/* Inlined functions for non-ABI compat mode */
-#include <odp/api/plat/queue_inlines.h>
+#include <odp/api/abi/queue_types.h>
+
+#include <odp/api/spec/queue_types.h>
 
 #ifdef __cplusplus
 }
