+++ include/odp/api/spec/schedule.h
@@ -20,7 +20,7 @@ extern "C" {
 
 #include <odp/api/std_types.h>
 #include <odp/api/event.h>
-#include <odp/api/queue.h>
+#include <odp/api/queue_types.h>
 #include <odp/api/schedule_types.h>
 #include <odp/api/thrmask.h>
