+++ include/odp/arch/mips64-linux/odp/api/abi/queue_types.h
@@ -1,0 +1,7 @@
+/* Copyright (c) 2021, Nokia
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier:     BSD-3-Clause
+ */
+
+#include <odp/api/abi-default/queue_types.h>
