+++ platform/linux-generic/Makefile.am
@@ -72,6 +72,7 @@ odpapiabiarchinclude_HEADERS += \
 		  include-abi/odp/api/abi/packet_io.h \
 		  include-abi/odp/api/abi/pool.h \
 		  include-abi/odp/api/abi/queue.h \
+		  include-abi/odp/api/abi/queue_types.h \
 		  include-abi/odp/api/abi/rwlock.h \
 		  include-abi/odp/api/abi/rwlock_recursive.h \
 		  include-abi/odp/api/abi/schedule.h \
