+++ platform/linux-generic/include/odp/api/plat/queue_inline_types.h
@@ -12,7 +12,7 @@ extern "C" {
 #endif
 
 #include <stdint.h>
-#include <odp/api/spec/queue_types.h>
+#include <odp/api/queue_types.h>
 
 /** @cond _ODP_HIDE_FROM_DOXYGEN_ */
